Initialization and setup
I have created the intial files and trying to learn how to use this thing called vi because I am too lazy to change it. Anyhow, pain is gain, and besides I need this working soon because these numbers need someone to deal with them. My current plan to write this out is none (a lot of regex), which I highly doubt can be ever be gotten to a working state. However, an alternative may present itself in the future, in which case I can dump this regex plan. Seriously, this thing is less regulated than a tax haven, and I have zero ways of knowing what are the boundaries when it comes to this dark magic. So as it stands it is vital to think of another plan and minimize the use of regex for my own sanity, I don't have too much to spare after dealing with this vi thing.